requirements.txt
agent-framework==1.0.0b251216
azure-cosmos>=4.5.0
python-dotenv>=1.0.0
httpx>=0.24.0